refactor(core): move non-component exports out of component files
openPresenterWindow, PRESET_COLORS, and the deck-warmth registry move
to their own modules so Fast Refresh can preserve state in player,
icon-picker, and slide-preload-layer instead of full-reloading. Deck
warmth also survives HMR edits to the preload layer now.

react-doctor: only-export-components
