+++ packages/core/src/app/components/sidebar/icon-picker.tsx
@@ -2,17 +2,7 @@ import EmojiPicker, { EmojiStyle, Theme } from 'emoji-picker-react';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import type { FolderIcon } from '@/lib/sdk';
 import { useLocale } from '@/lib/use-locale';
-
-export const PRESET_COLORS = [
-  '#e5484d', // red
-  '#f76b15', // orange
-  '#ffb224', // amber
-  '#30a46c', // green
-  '#12a594', // teal
-  '#0091ff', // blue
-  '#6e56cf', // violet
-  '#60646c', // gray
-];
+import { PRESET_COLORS } from './preset-colors';
 
 export function IconPicker({
   value,
