+++ packages/core/src/app/components/sidebar/preset-colors.ts
@@ -1,0 +1,10 @@
+export const PRESET_COLORS = [
+  '#e5484d', // red
+  '#f76b15', // orange
+  '#ffb224', // amber
+  '#30a46c', // green
+  '#12a594', // teal
+  '#0091ff', // blue
+  '#6e56cf', // violet
+  '#60646c', // gray
+];
